test Ref/RefMut protector interactions
